chore: remove duplicates and consolidate ICP tool
