+++ config/ai-mapping.json
@@ -6,7 +6,6 @@
   "content": "claude-3.5-sonnet",
   "usp": "deepseek-v3",
   "icp": "deepseek-v3",
-  "icp-maker": "deepseek-v3",
   "offer-generator": "grok-3",
   "title-generator": "deepseek-v3",
   "description-generator": "deepseek-v3",
