+++ src/app/api/icp/generate-v2/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '@/lib/auth';
 import { openai } from '@/lib/openai';
-import { ICPFormData } from '@/app/outils/icp-maker/page';
+import { ICPFormData } from '@/types/icp';
 
 export const dynamic = 'force-dynamic';
